+++ tests/Phpunit/Unit/Domain/Model/AuditBudgetTest.php
@@ -183,6 +183,26 @@ final class AuditBudgetTest extends TestCase
         AuditBudget::forBoth(100, \NAN);
     }
 
+    /**
+     * @throws InvalidAuditBudgetException
+     */
+    public function test_for_cost_rejects_infinite(): void
+    {
+        $this->expectException(InvalidAuditBudgetException::class);
+
+        AuditBudget::forCost(\INF);
+    }
+
+    /**
+     * @throws InvalidAuditBudgetException
+     */
+    public function test_for_both_rejects_infinite_cost(): void
+    {
+        $this->expectException(InvalidAuditBudgetException::class);
+
+        AuditBudget::forBoth(100, \INF);
+    }
+
     public function test_for_cost_rejects_negative_and_formats_the_value_with_a_period_regardless_of_the_process_numeric_locale(): void
     {
         $previousLocale = setlocale(\LC_NUMERIC, '0');